pyproject.toml
[project]
name = "digital-krishi-officer"
version = "0.1.0"
description = "Digital Krishi Officer Backend API"
authors = [{name = "SIH Team", email = "team@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "openai>=1.107.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "redis>=6.4.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
    "aiofiles>=24.1.0",
    "requests>=2.32.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
